+++ includes/functions/parser.h
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 20:41:18 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/15 23:02:02 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/03/28 22:21:45 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@ t_item		*line2cylinder(char **line);
 void		get_items(t_itemlist **list,char *argv);
 t_corners	set_corners(t_item item);
 t_coords	**set_cam_vectors(t_corners corn);
+void		free_array_coords(t_coords ***list);
 
 
 #endif